Execute MySQL operations on the remote server.
1/ Port forwarding ended up being too slow, in particular due to TCP on TCP
performance when running MySQL commands through a SSH tunnel.

2/ Redefining mysql, mysqlshow, and mysqldump as bash functions that remote
execute over ssh would have worked except that bash passes parsed arguments to
function (there is no practical way to get the raw string that follows the
function call), which removes important double-quotes.

3/ This left remote execution of commands and dealing locally with the output
as the most viable option, even if this meant more modifications of the script
(meaning that it will be more time consuming to upgrade in the future to newer
versions of automysqlbackup).